Import latest Samsung release
* Package version: T713XXU2BQCO

+++ arch/arm64/include/asm/processor.h
@@ -140,7 +140,7 @@ extern struct task_struct *cpu_switch_to(struct task_struct *prev,
 #define task_pt_regs(p) \
 	((struct pt_regs *)(THREAD_START_SP + task_stack_page(p)) - 1)
 
-#define KSTK_EIP(tsk)	task_pt_regs(tsk)->pc
+#define KSTK_EIP(tsk)	((unsigned long)task_pt_regs(tsk)->pc)
 #ifndef CONFIG_COMPAT
 #define KSTK_ESP(tsk)	task_pt_regs(tsk)->sp
 #else
